extract module parser common helpers
