+++ test/pr476_parse_module_common_helpers.test.ts
@@ -1,0 +1,49 @@
+import { describe, expect, it } from 'vitest';
+
+import type { Diagnostic } from '../src/diagnostics/types.js';
+import {
+  consumeKeywordPrefix,
+  parseReturnRegsFromText,
+  parseVarDeclLine,
+  topLevelStartKeyword,
+} from '../src/frontend/parseModuleCommon.js';
+import { makeSourceFile, span } from '../src/frontend/source.js';
+
+describe('PR476 module helper extraction', () => {
+  const file = makeSourceFile('pr476_parse_module_common_helpers.zax', '');
+  const zeroSpan = span(file, 0, 0);
+
+  it('keeps top-level keyword helpers intact', () => {
+    expect(consumeKeywordPrefix('export func main()', 'export')).toBe('func main()');
+    expect(topLevelStartKeyword('export const FOO = 1')).toBe('const');
+    expect(topLevelStartKeyword('nonsense thing')).toBeUndefined();
+  });
+
+  it('keeps return register parsing intact', () => {
+    const diagnostics: Diagnostic[] = [];
+    expect(parseReturnRegsFromText('HL, de', zeroSpan, 1, diagnostics, file.path)).toEqual({
+      regs: ['HL', 'DE'],
+    });
+    expect(diagnostics).toEqual([]);
+  });
+
+  it('keeps var declaration parsing intact', () => {
+    const diagnostics: Diagnostic[] = [];
+    const decl = parseVarDeclLine('value: word = $12', zeroSpan, 1, 'var', {
+      diagnostics,
+      modulePath: file.path,
+      isReservedTopLevelName: () => false,
+    });
+
+    expect(diagnostics).toEqual([]);
+    expect(decl).toMatchObject({
+      kind: 'VarDecl',
+      name: 'value',
+      typeExpr: { kind: 'TypeName', name: 'word' },
+      initializer: {
+        kind: 'VarInitValue',
+        expr: { kind: 'ImmLiteral', value: 0x12 },
+      },
+    });
+  });
+});
